+++ Changelog.md
@@ -1,5 +1,17 @@
 ## 0.0.71
-*INCOMPLETE*
++ Added `beta.openai.Annotations.FrequencyPenalty` annotation method to specify LLM frequency penalty parameters.
++ Added `beta.openai.Annotations.MaxOutputTokens` annotation method to specify LLM maximum output token parameters.
++ Added `beta.openai.Annotations.PresencePenalty` annotation method to specify LLM presence penalty parameters.
++ Added `beta.openai.Annotations.ReasoningEffort` annotation method to specify LLM reasoning effort parameters.
++ Added `beta.openai.Annotations.ResponseFormat` annotation method to specify LLM response format parameters.
++ Added `beta.openai.Annotations.ResponseStreaming` annotation method to specify LLM response streaming parameters.
++ Added `beta.openai.Annotations.SamplingProbability` annotation method to specify LLM sampling probability parameters.
++ Refactored `Parameter.AudioSpeed` annotation method to `beta.openai.Annotations.AudioSpeed`.
++ Refactored `Parameter.AudioVoice` annotation method to `beta.openai.Annotations.AudioVoice`.
++ Refactored `Parameter.EmbeddingDims` annotation method to `beta.openai.Annotations.EmbeddingDims`.
++ Refactored `Parameter.SamplingTemperature` annotation method to `beta.openai.Annotations.SamplingTemperature`.
++ Refactored `Parameter.TranscriptionLanguage` annotation method to `beta.openai.Annotations.TranscriptionLanguage`.
++ Refactored `Parameter.TranscriptionPrompt` annotation method to `beta.openai.Annotations.TranscriptionPrompt`.
 
 ## 0.0.70
 + Added `muna.beta.openai.audio.speech.create` method for using text-to-speech predictors via the OpenAI API.