+++ Changelog.md
@@ -1,3 +1,6 @@
+## 0.0.72
+*INCOMPLETE*
+
 ## 0.0.71
 + Added `beta.openai.Annotations.FrequencyPenalty` annotation method to specify LLM frequency penalty parameters.
 + Added `beta.openai.Annotations.MaxOutputTokens` annotation method to specify LLM maximum output token parameters.